+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityEggPhysicsComponentImpl.java
@@ -3,7 +3,6 @@ package org.allaymc.server.entity.component.projectile;
 import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.EntityInitInfo;
-import org.allaymc.api.entity.component.EntityPhysicsComponent;
 import org.allaymc.api.entity.damage.DamageContainer;
 import org.allaymc.api.entity.interfaces.EntityLiving;
 import org.allaymc.api.entity.type.EntityTypes;
@@ -55,11 +54,9 @@ public class EntityEggPhysicsComponentImpl extends EntityProjectilePhysicsCompon
         if (other instanceof EntityLiving living) {
             // Eggs deal 0 damage but still have knockback
             var damage = DamageContainer.projectile(thisEntity, 0);
-            damage.setHasKnockback(false);
-            if (living.attack(damage) && other instanceof EntityPhysicsComponent physicsComponent) {
-                // Use the last location as the knockback source
-                physicsComponent.knockback(hitPos.sub(this.motion, new Vector3d()));
-            }
+            // Use the last location as the knockback source
+            damage.setKnockbackSource(hitPos.sub(this.motion, new Vector3d()));
+            living.attack(damage);
         }
 
         handleEggBreak(hitPos);